+++ src/Circle.h
@@ -4,9 +4,9 @@
 class Circle
 {
 public:
-    Circle(int radius_, int circleWidth_, ofVec3f center_);
+    Circle(int radius_, float circleWidth_, ofVec3f center_);
     virtual ~Circle();
-    void update(ofVec3f center_, int circleWidth_,int radius_);
+    void update(ofVec3f center_, float circleWidth_,int radius_);
     void setRadius(int radius_);
     int getRadius();
     int getCircleWidth();
